% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ps.R
\name{head.ps}
\alias{head.ps}
\title{returns the head of the pareto set}
\usage{
\method{head}{ps}(x, n = 6L, ...)
}
\arguments{
\item{x}{an object}

\item{n}{a single integer. If positive, size for the resulting
    object: number of elements for a vector (including lists), rows for
    a matrix or data frame or lines for a function. If negative, all but
    the \code{n} last/first number of elements of \code{x}.}

\item{...}{arguments to be passed to or from other methods.}
}
\description{
Description
}
\examples{
aps <- ps(matrix(rnorm(1:1000),ncol=2))
head(aps)

}

