% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mco2ps}
\alias{mco2ps}
\title{Converts 'mco' objects into 'ps' objects}
\usage{
mco2ps(mco)
}
\arguments{
\item{mco}{a 'mco' object}
}
\value{
a 'ps' object
}
\description{
Converts 'mco' objects into 'ps' objects
}

