% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moko.R
\docType{package}
\name{moko}
\alias{moko}
\title{moko: Multi-objective Kriging Optimization}
\description{
The package \code{moko} provides the user with methods for constrained and unconstrained
multi-objective optimization based on the popular Kriging surrogate model.
}
\details{
The main functions provided by \code{moko} are:
\code{\link{MEGO}}, \code{\link{HEGO}} and \code{\link{MVPF}}.
}
