% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_functions.R
\docType{data}
\name{nowacki_beam_tps}
\alias{nowacki_beam_tps}
\title{True pareto front for the nowacki beam problem}
\format{An object of class \code{ps} of length 4.}
\usage{
nowacki_beam_tps
}
\description{
True pareto front for the nowacki beam problem
}
\keyword{datasets}
