% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ps.R
\name{tail.ps}
\alias{tail.ps}
\title{returns the tail of the pareto set}
\usage{
\method{tail}{ps}(x, n = 6L, addrownums = TRUE, ...)
}
\arguments{
\item{x}{an object}

\item{n}{a single integer. If positive, size for the resulting
    object: number of elements for a vector (including lists), rows for
    a matrix or data frame or lines for a function. If negative, all but
    the \code{n} last/first number of elements of \code{x}.}

\item{addrownums}{if there are no row names, create them from the row
    numbers.}

\item{...}{arguments to be passed to or from other methods.}
}
\description{
Description
}
\examples{
aps <- ps(matrix(rnorm(1:1000),ncol=2))
tail(aps)

}

