% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optim_hego.R
\name{EHVI}
\alias{EHVI}
\title{Title}
\usage{
EHVI(x, model, control = NULL)
}
\arguments{
\item{x}{a vector representing the input for which one wishes to calculate EHI}

\item{model}{object of class \code{mkm} containing the objectives and constrains}

\item{paretoFront}{optional object of class \code{ps} containing the actual Pareto set}

\item{minimization}{logical indicating if the EHVI is minimizing all objectives
(\code{TRUE} by default) or maximizing all objectives (\code{FALSE})}
}
\value{
The Constrained Expected Hypervolume Improvement at \code{x}.
}
\description{
Description
}
\examples{
# ------------------------
# The Nowacki Beam
# ------------------------
n <- 50
d <- 2
doe <- replicate(d,sample(0:n,n))/n
res <- t(apply(doe, 1, nowacki_beam))
model <- mkm(doe, res, modelcontrol = list(objective = 1:2, lower=rep(0.1,d)))
grid <- expand.grid(seq(0, 1, , 50),seq(0, 1, , 50))
ehvi <- apply(grid, 1, EHVI, model) # this computation may take some time
contour(matrix(ehvi, 50))
points(model@design, col=ifelse(model@feasible,'blue','red'))
points(grid[which.max(ehvi),], col='green', pch=19)
}

