% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optim_pego.R
\name{EI}
\alias{EI}
\title{Constrained expected improvement}
\usage{
EI(x, model, control = NULL)
}
\description{
description
}
\examples{
# --------------------
# Branin-Hoo function
# --------------------
n <- 20
d <- 2
doe <- replicate(d,sample(0:n,n))/n
fun_cost <- DiceKriging::branin
fun_cntr <- function(x) 0.2 - prod(x)
fun <- function(x) return(cbind(fun_cost(x),fun_cntr(x)))
res <- t(apply(doe, 1, fun))
model <- mkm(doe, res, modelcontrol = list(objective = 1, lower=c(0.1,0.1)))
grid <- expand.grid(seq(0,1,,50),seq(0,1,,50))
ei <- apply(grid, 1, EI, model) # this computation may take some time
contour(matrix(ei,50))
points(model@design, col=ifelse(model@feasible,'blue','red'))
# ------------------------
# The Nowacki Beam
# ------------------------
fun_cost <- DiceKriging::branin
fun_cntr <- function(x){
 g1 <- 0.9 - sum(x)
 g2 <- sum(x) - 1.1
 g3 <- - x[1] + 0.5
 g4 <- x[2] - 0.5
 return(c(g1,g2,g3,g4))
}
fun <- function(x) return(c(fun_cost(x),fun_cntr(x)))
fun <- nowacki_beam
n <- 50
d <- 2
plugin <- NULL
doe <- replicate(d,sample(0:n,n))/n
doe <- replicate(d,sample(5:n,n-5))/n
for (i in 1:80){
res <- t(apply(doe, 1, fun))
res <- cbind(moko:::Tchebycheff(res[,1:2]),res[,-(1:2)])
#res <- cbind(res[,2],res[,-(1:2)])
model <- mkm(doe, res, modelcontrol = list(objective = 1, lower=rep(0.1,d)))
grid <- expand.grid(seq(0.1, 1, , 25),seq(0.1, 1, , 25))
ei <- apply(grid, 1, EI, model)
#, control=list(plugin=NULL)) # this computation may take some time
contour(matrix(ei, 25))
points(model@design, col=ifelse(model@feasible,'blue','red'))
points(grid[which.max(ei),], col='green', pch=19)
x_star <- unlist(grid[which.max(ei),])
y_star <- fun(x_star)
doe <- rbind(doe, x_star)
row.names(doe) <- NULL
}
}

