% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optim_hego.R
\name{HEGO}
\alias{HEGO}
\title{HEGO: Efficient Global Optimization Algorithm based on the Hypervolume criteria}
\usage{
HEGO(model, fun, nsteps, lower = rep(0, model@d), upper = rep(1, model@d),
  quiet = TRUE, control = NULL, optimcontrol = NULL)
}
\arguments{
\item{model}{An object of class \code{\link{mkm}}.}

\item{fun}{The multi-objective and constraint cost function to be optimized.
This function must return a vector with the size of \code{model@m + model@j}
where \code{model@m} are the number of objectives and
\code{model@j} the number of the constraints,}

\item{nsteps}{An integer representing the desired number of iterations,}

\item{lower}{Vector of lower bounds for the variables to be optimized over
(default: 0 with length \code{model@d}),}

\item{upper}{Vector of upper bounds for the variables to be optimized over
(default: 1 with length \code{model@d}),}

\item{quiet}{Logical indicating the verbosity of the routine,}

\item{control}{An optional list of control parameters, some of them passed to
the \code{\link[GPareto]{crit_EHI}} function. One can control:
\describe{
\item{\code{minimization}}{logical indicating if the EHVI is minimizing
all objectives (\code{TRUE}, by default) or maximizing all objectives
(\code{FALSE}). Mixed information not currently accepted, if the user needs
so, it should invert the functions prior Kriging modeling}
\item{\code{paretoFront}}{object of class \code{\link{ps}} containing
the actual Pareto set. If not provided a Pareto set is built based
 on the current observations of \code{model}.}
\item{\code{nb.samp}}{default: 50}
\item{\code{seed}}{default: 42}
\item{\code{refPoint}}{default: \code{min} or \code{max} values for the Pareto
 front for each objective}
}}

\item{optimcontrol}{Optional list of control parameters passed to the
\code{\link[GenSA]{GenSA}} function. Please, note that the values are passed as
the \code{control} parameter inside the \code{GenSA} function.}
}
\value{
updated \code{\link{mkm}} model
}
\description{
Executes \code{nsteps} iterations of the HEGO method to an object of class
\code{\link{mkm}}. At each step, a kriging model is re-estimated (including
covariance parameters re-estimation) based on the initial design points plus
the points visited during all previous iterations; then a new point is
obtained by maximizing the Expected Hypervolume Improvement criterion (EHVI).
}
\examples{
# ----------------
# Fonseca and Flemming
# ----------------
n <- 20
d <- 2
m <- 2
A <- 4
fun <- Fonseca
doe <- 2*A*replicate(d,sample(0:n,n))/n - A
res <- t(apply(doe, 1, fun))
model <- mkm(doe, res, modelcontrol=list(lower=rep(0.1,d)))
model <- HEGO(model, fun, 20, lower = -rep(A,d), upper = rep(A,d), quiet = FALSE)
tpf <- mco::nsga2(fun, d, 2, lower.bounds = -rep(A,d), upper.bounds = rep(A,d))$value
plot(tpf)
points(ps(model@response)$set, col = 'blue', pch = 19)
# ----------------
# Shaffer1
# ----------------
n <- 10
d <- 1
A <- 10
fun <- Shaffer1
doe <- 2*A*replicate(d,sample(0:n,n))/n - A
res <- t(apply(doe, 1, fun))
model <- mkm(doe, res, modelcontrol=list(lower=rep(0.1,d)))
model <- HEGO(model, fun, 20, lower = -A, upper = A, quiet = FALSE)
tpf <- mco::nsga2(fun, d, 2, lower.bounds = -A, upper.bounds = A)$value
plot(tpf)
points(ps(model@response)$set, col = 'blue', pch = 19)
# ----------------
# Shaffer2
# ----------------
n <- 10
d <- 1
fun <- Shaffer2
doe <- 15 * replicate(d,sample(0:n,n))/n - 5
res <- t(apply(doe, 1, fun))
model <- mkm(doe, res, modelcontrol=list(lower=rep(0.1,d)))
model <- HEGO(model, fun, 20, lower = -5, upper = 10, quiet = FALSE)
tpf <- mco::nsga2(fun, d, 2, lower.bounds = -5, upper.bounds = 10)$value
plot(tpf)
points(ps(model@response)$set, col = 'blue', pch = 19)
# ----------------
# Viennet
# ----------------
n <- 20
d <- 2
fun <- Viennet
doe <- replicate(d,sample(0:n,n))/n
res <- t(apply(doe, 1, fun))
model <- mkm(doe, res, modelcontrol=list(lower=rep(0.1,d)))
model <- HEGO(model, fun, 80, quiet = FALSE)
pairs(ps(model@response)$set)
rgl::plot3d(ps(model@response)$set)
# ----------------
# Binh
# ----------------
n <- 20
d <- 2
fun <- Binh
doe <- cbind(rep(5,n), rep(3,n)) * replicate(d,sample(0:n,n))/n
res <- t(apply(doe, 1, fun))
model <- mkm(doe, res, modelcontrol = list(lower=rep(0.1,d), objectives = 1:2))
model <- HEGO(model, fun, 40, upper = c(5,3), quiet = FALSE)
fun <- function(x) Binh(x)[c(1,2)]
gfun <- function(x) -Binh(x)[-c(1,2)]
tpf <- mco::nsga2(fun, d, 2, lower.bounds = c(0,0), upper.bounds = c(5,3),
                   constraints = gfun, cdim = 2)$value
plot(tpf)
points(ps(model@response[which(model@feasible),model@objective])$set, col = 'blue', pch = 19)
# ----------------
# The Nowacki Beam
# ----------------
n <- 20
d <- 2
fun <- nowacki_beam
doe <- replicate(d,sample(0:n,n))/n
res <- t(apply(doe, 1, fun))
model <- mkm(doe, res, modelcontrol = list(objective = 1:2, lower = rep(0.1,d)))
model <- HEGO(model, fun, 80, quiet = FALSE, control = list(rho = 0.1))
fun <- function(x) nowacki_beam(x)[c(1,2)]
gfun <- function(x) -nowacki_beam(x)[-c(1,2)]
tpf <- mco::nsga2(fun, d, 2, lower.bounds = c(0,0), upper.bounds = c(1,1),
                   constraints = gfun, cdim = 4)$value
plot(tpf)
points(ps(model@response[which(model@feasible),model@objective])$set, col = 'blue', pch = 19)
}

