% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ps_metrics.R
\name{igd}
\alias{igd}
\title{IGD pareto quality metric}
\usage{
igd(aps, tps, method = "manhattan", norm = TRUE)
}
\arguments{
\item{method}{the distance measure to be used. This must be one of "euclidean" or "manhattan" (default)}

\item{apf}{'ps' object containing the 'actual' pareto front}

\item{tpf}{'ps' object containing the 'true' pareto front}
}
\value{
returns the IGD metric
}
\description{
This function implements the method described in shimoyama2013kriging
}
\examples{
aps <- ps(matrix(rnorm(1:1000),ncol=2))
tps <- ps(matrix(rnorm(1:2000),ncol=2))
igd(aps,tps)
}

