% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optim_pego.R
\name{max_EI}
\alias{max_EI}
\title{Title}
\usage{
max_EI(model, lower = rep(0, model@d), upper = rep(1, model@d),
  control = NULL, optimcontrol = list(max.time = 2))
}
\arguments{
\item{model}{An object of class \code{\link{mkm}}.}

\item{lower}{
        Vector with length of \code{par}. Lower bounds for components.
    }

\item{upper}{
        Vector with length of \code{par}. Upper bounds for components.
    }

\item{control}{
        The argument is a list that can be used to control the behavior of the algorithm:
            \describe{
                \item{\code{maxit}}{
                    Integer. Maximum number of iterations of the algorithm. 
                }
                \item{\code{threshold.stop}}{
                    Numeric. The program will stop when the expected objective function value \code{threshold.stop} is reached. Default value is \code{NULL}
                }
                \item{\code{nb.stop.improvement}}{
                    Integer. The program will stop when there is no any improvement in \code{nb.stop.improvement} steps. 
                }
                \item{\code{smooth}}{
                    Logical.\code{TRUE} when the objective function is smooth, or differentiable almost everywhere in the
                        region of \code{par}, \code{FALSE} otherwise. Default value is \code{TRUE}.
                }
                \item{\code{max.call}}{
                    Integer. Maximum number of call of the objective function. Default is set to 1e7.
                }
                \item{\code{max.time}}{
                    Numeric. Maximum running time in seconds.
                }
                \item{\code{temperature}}{
                    Numeric. Initial value for temperature.
                }
                \item{\code{visiting.param}}{
                    Numeric. Parameter for visiting distribution.
                }
                \item{\code{acceptance.param}}{
                    Numeric. Parameter for acceptance distribution.
                }
                \item{\code{verbose}}{
                    Logical. \code{TRUE} means that messages from the algorithm are shown. Default is \code{FALSE}.
                }
                \item{\code{simple.function}}{
                    Logical. \code{FALSE} means that the objective function has only a few local minima. Default is \code{FALSE} which means that the objective function is complicated with many local minima.
                }
                \item{\code{trace.mat}}{
                    Logical. Default is \code{TRUE} which means that the trace matrix will be available in the returned value of \code{GenSA} call.
                }
            }
    }
}
\value{
Vector. The best set of parameters found.
}
\description{
Description
}
\examples{
# --------------------
# Branin-Hoo function
# --------------------
n <- 20
d <- 2
doe <- replicate(d,sample(0:n,n))/n
res <- apply(doe, 1, DiceKriging::branin)
model <- mkm(doe, res)
max_EI(model)
}

