% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mkm.R
\name{mkm}
\alias{mkm}
\title{Multiobjective Kriging model}
\usage{
mkm(design, response, modelcontrol = NULL)
}
\arguments{
\item{design}{ a data frame representing the design of experiments. The ith row contains the values of the d input variables corresponding to the ith evaluation}

\item{response}{ a vector (or 1-column matrix or data frame) containing the values of the 1-dimensional output given by the objective function at the \code{design} points. }

\item{modelcontrol}{list of parameters passed to \code{\link{km}}.}

\item{quiet}{logical indicating if the \code{print} commands inside \code{km} should be ommited}
}
\value{
S4 class object that contains a list of \code{\link{km}} objects and also
 some usefull design data
}
\description{
This function creates a multiobjective kriging model. It is based on the \code{\link{km}}
 function of the \code{\link{DiceKriging}} package and creates a structured list of
 \code{\link{km}} objects.
}
\examples{
# ------------------------
# The Nowacki Beam
# ------------------------
n <- 10
d <- 2
doe <- replicate(d,sample(0:n,n))/n
res <- t(apply(doe, 1, nowacki_beam))
model <- mkm(doe, res, modelcontrol = list(objective = 1:2))
}

