% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moko.R
\docType{package}
\name{moko}
\alias{moko}
\alias{moko-package}
\title{moko: Multiobjective Kriging Optimization}
\description{
The package \code{moko} provides the user with methods for constrained and unconstraiend
multiobjective optimization.
}
\details{
The main functions provided by \code{moko} are:
pEGO, hEGO and VMFK
}

