% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_functions.R
\name{nowacki_beam}
\alias{nowacki_beam}
\title{Test function: The Nowacki Beam}
\usage{
nowacki_beam(x, g = c(5, 240, 120, 10), l = 1500, F = 5000, E = 216620,
  G = 86650, v = 0.27, box = data.frame(b = c(10, 50), h = c(20, 250)))
}
\arguments{
\item{x}{vector of length 2 correspon the normalized beath and height of the beam}

\item{g}{vector of lenght 4 containing the limit of each constraint}

\item{l}{numeric length of the beam}

\item{F}{numeric force applied at the beam tip}

\item{E}{numeric elastic longitudinal moduli}

\item{G}{numeric elastic transversal moduli}

\item{v}{numeric poison ratio}

\item{box}{data.frame structure containing the upper and lower limits for \code{b} and \code{h}}
}
\value{
vector of objective and constrain responses
}
\description{
Description
}
\examples{

nowacki_beam(c(0.5,0.5))
}

