% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ps_metrics.R
\name{pdist}
\alias{pdist}
\title{Distance betwen vector and matrix}
\usage{
pdist(point, set, method = "manhattan")
}
\arguments{
\item{point}{numeric vector}

\item{set}{numeric matrix}

\item{method}{the distance measure to be used. This must be one of "euclidean" or "manhattan" (default)}
}
\value{
numeric value indicating the minimum distance between point and set
}
\description{
This function computes and returns the minimum distance between a vector and a matrix
}
\details{
#@export
}
\examples{
aps <- ps(matrix(rnorm(1:1000),ncol=2))
pdist(c(0,0), aps$set)
}

